+++ Sources/ForTag.swift
@@ -42,7 +42,7 @@ class ForNode : NodeType {
     let resolvable = try parser.compileResolvable(components[3])
 
     let `where` = hasToken("where", at: 4)
-      ? try parseExpression(components: Array(components.suffix(from: 5)), tokenParser: parser)
+      ? try parser.compileExpression(components: Array(components.suffix(from: 5)))
       : nil
 
     return ForNode(resolvable: resolvable, loopVariables: loopVariables, nodes: forNodes, emptyNodes:emptyNodes, where: `where`)